requirements.txt
aerich==0.7.2
aiofiles==24.1.0
aiogram
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
executing==2.1.0
fastadmin==0.2.13
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
icecream==2.1.3
idna==3.10
iso8601==1.1.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.1
multidict==6.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomlkit==0.13.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
unicorn==2.1.1
uvicorn==0.31.0
yarl==1.14.0